MONDRIAN: Check XML jars into lib directory; no longer need XALAN_HOME to be set.
[git-p4: depot-paths = "//open/mondrian/": change = 3542]

+++ testsrc/main/mondrian/test/loader/MondrianFoodMartLoader.java
@@ -773,7 +773,8 @@ public class MondrianFoodMartLoader {
     {
         try {
             StringBuffer buf = new StringBuffer();
-            if (jdbcOutput) {
+            // If we're [re]creating tables, no need to drop indexes.
+            if (jdbcOutput && !tables) {
                 try {
                     buf.append("DROP INDEX ")
                         .append(quoteId(indexName));
